+++ examples/src/data_model_examples.cpp
@@ -13,7 +13,7 @@ void data_model_example1()
     j.emplace_back("foo");
     j.emplace_back(byte_string{ 'b','a','r' });
     j.emplace_back("-18446744073709551617", semantic_tag_type::bignum);
-    j.emplace_back("273.15", semantic_tag_type::decimal);
+    j.emplace_back("273.15", semantic_tag_type::decimal_fraction);
     j.emplace_back("2018-10-19 12:41:07-07:00", semantic_tag_type::date_time);
     j.emplace_back(1431027667, semantic_tag_type::epoch_time);
     j.emplace_back(-1431027667, semantic_tag_type::epoch_time);